RTDirRead/RTDirReadEx: sizes are size_t not unsigned.
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@14062 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Devices/Storage/VBoxHDD-new.cpp
@@ -690,7 +690,7 @@ VBOXDDU_DECL(int) VDBackendInfo(unsigned cEntriesAlloc, PVDBACKENDINFO pEntries,
             break;
 
         PRTDIRENTRYEX pPluginDirEntry = NULL;
-        unsigned cbPluginDirEntry = sizeof(RTDIRENTRY);
+        size_t cbPluginDirEntry = sizeof(RTDIRENTRY);
         pPluginDirEntry = (PRTDIRENTRYEX)RTMemAllocZ(sizeof(RTDIRENTRY));
         if (!pPluginDirEntry)
         {
@@ -984,7 +984,7 @@ VBOXDDU_DECL(int) VDGetFormat(const char *pszFilename, char **ppszFormat)
         }
 
         PRTDIRENTRYEX pPluginDirEntry = NULL;
-        unsigned cbPluginDirEntry = sizeof(RTDIRENTRY);
+        size_t cbPluginDirEntry = sizeof(RTDIRENTRY);
         pPluginDirEntry = (PRTDIRENTRYEX)RTMemAllocZ(sizeof(RTDIRENTRY));
         if (!pPluginDirEntry)
         {
